wsc/WochuServerCenter/requirements.txt
astroid==1.6.3
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
dateutils==0.6.6
Django==2.0.4
django-braces==1.12.0
django-cors-headers==2.2.0
django-filter==1.1.0
django-oauth-toolkit==1.1.0
django-redis-sessions==0.6.1
django-rest-framework-docs==0.1.7
django-rest-swagger==2.1.0
django-silk==3.0.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-dynamic-fields==0.3.0
gprof2dot==2016.10.13
httpie==0.9.9
idna==2.6
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.7
Pygments==2.2.0
PyJWT==1.6.1
pylint==1.8.4
PyMySQL==0.8.0
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
wrapt==1.10.11
xlwt==1.3.0
oss2==2.5.0
django-extensions==2.1.2
uWSGI==2.0.17.1
